third_party/blink/web_tests/accessibility/computed-text-with-height-0.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div class="container">
    <a id="button" href="#" role="button"><span style="height:0px;position:absolute;overflow:hidden">Friend Requests:</span><span>0</span></a>
</div>

<script>
test(function(t){
    var axButton = accessibilityController.accessibleElementById("button");
    assert_equals(axButton.name, "Friend Requests: 0");
    assert_equals(axButton.name, "Friend Requests: 0");
}, "Text with height:0, position:absolute, and overflow:hidden should still be included in the accessible text computation.");
</script>
